move partition_for_dnnl from dnnl.py to test_dnnl.py
